requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
